Rename package to avoid lib-go confusion
