+++ pkg/tnf/pkg/jobs/batch.go
@@ -1,4 +1,4 @@
-package library_go
+package jobs
 
 import (
 	"context"
