+++ pkg/tnf/pkg/jobs/jobcontroller.go
@@ -1,4 +1,4 @@
-package library_go
+package jobs
 
 import (
 	"context"
